Refactor(modules) code to improve readability and maintainability
- Updated index.html for better structure
- Improved functions in js/index.js
- Refactored Alert.js and Appointment.js classes
- Enhanced utility functions in js/modules/functions.js
- Cleaned up selectors in js/modules/selectors.js
- Organized variables in js/modules/variables.js
