+++ index.html
@@ -35,5 +35,5 @@
         </div>
     </div>
 </body>
-<script src="./js/index.js"></script>
+<script src="./js/index.js" type="module"></script>
 </html>
